pyproject.toml
[project]
name = "isolysis"
version = "0.7.0"
description = "Isochrone computation API with spatial analysis capabilities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alphashape>=1.3.1",
    "contextily>=1.6.2",
    "fastapi[standard]>=0.116.1",
    "folium>=0.20.0",
    "geopandas>=1.1.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.5",
    "openpyxl>=3.1.5",
    "osmnx>=2.0.5",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rasterio>=1.4.3",
    "rasterstats>=0.20.0",
    "routingpy>=1.3.0",
    "scikit-learn>=1.7.1",
    "streamlit>=1.40.0",
    "streamlit-folium>=0.25.1",
]

[tool.setuptools.packages.find]
include = ["isolysis*"]
exclude = [
  "api",
  "data",
  "outputs",
  "networks",
  "research",
  "cache",
  "tests*",
  "scripts*"
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "leafmap>=0.49.0",
    "notebook>=7.4.5",
    "pytest>=8.4.1",
]
